+++ src/lib.rs
@@ -14,6 +14,15 @@
 //! distribution it runs on, follows this specification.
 //!
 
+#![deny(
+    rust_2018_compatibility,
+    rust_2018_idioms,
+    nonstandard_style,
+    unused,
+    future_incompatible,
+    clippy::all
+)]
+
 use std::ffi::OsString;
 use std::hash::{Hash, Hasher};
 use std::path::{Path, PathBuf};
@@ -39,7 +48,7 @@ pub mod macos;
 #[cfg(target_os = "macos")]
 use macos as platform;
 
-// pub use platform as my_latform;
+// pub use platform as my_platform;
 pub const DEFAULT_TRASH_CTX: TrashContext = TrashContext::new();
 
 /// A collection of preferences for trash operations.
@@ -54,7 +63,7 @@ impl TrashContext {
 
     /// Removes a single file or directory.
     ///
-    /// When a symbolic link is provided to this function, the sybolic link will be removed and the link
+    /// When a symbolic link is provided to this function, the symbolic link will be removed and the link
     /// target will be kept intact.
     ///
     /// # Example
@@ -72,7 +81,7 @@ impl TrashContext {
 
     /// Removes all files/directories specified by the collection of paths provided as an argument.
     ///
-    /// When a symbolic link is provided to this function, the sybolic link will be removed and the link
+    /// When a symbolic link is provided to this function, the symbolic link will be removed and the link
     /// target will be kept intact.
     ///
     /// # Example
@@ -224,10 +233,10 @@ where
 /// This struct holds information about a single item within the trash.
 ///
 /// Some functions associated with this struct are defined in the
-/// `TrahsItemPlatformDep` trait. That trait is implemented for `TrashItem` by
+/// `TrashItemPlatformDep` trait. That trait is implemented for `TrashItem` by
 /// each platform specific source file individually.
 ///
-/// A trahs item can be a file or folder or any other object that the target
+/// A trash item can be a file or folder or any other object that the target
 /// operating system allows to put into the trash.
 #[derive(Debug)]
 pub struct TrashItem {
@@ -259,7 +268,7 @@ pub struct TrashItem {
 
 /// Platform independent functions of `TrashItem`.
 ///
-/// See `TrahsItemPlatformDep` for platform dependent functions.
+/// See `TrashItemPlatformDep` for platform dependent functions.
 impl TrashItem {
     /// Joins the `original_parent` and `name` fields to obtain the full path to
     /// the original file.